Casting as long instead of int for digital bitwise stuff

+++ src/symphony-core/HekaDAQController.Tests/HekaDAQStreamTests.cs
@@ -128,8 +128,8 @@ namespace Heka
                 s.BitPositions[dev] = bitPosition;
             }
 
-            ushort q = 0xaaaa;
-            var expected = new Measurement((short)q, Measurement.UNITLESS);
+            ulong q = 0xaaaa;
+            var expected = new Measurement((long)q, Measurement.UNITLESS);
 
             Assert.AreEqual(expected, s.Background);
         }
@@ -159,7 +159,7 @@ namespace Heka
                 s.BitPositions[dev] = bitPosition;
             }
 
-            var expected = Enumerable.Range(0, 10000).Select(i => new Measurement((short)(i % 2 * 0xaaaa), Measurement.UNITLESS)).ToList();
+            var expected = Enumerable.Range(0, 10000).Select(i => new Measurement((long)(i % 2 * 0xaaaa), Measurement.UNITLESS)).ToList();
 
             var pull1 = s.PullOutputData(duration);
             Assert.AreEqual(expected, pull1.Data);
