+++ src/symphony-core/HekaDAQController/HekaDAQInputStream.cs
@@ -129,7 +129,7 @@ namespace Heka
                 data = new InputData(data,
                                      data.Data.Select(
                                          m =>
-                                         MeasurementPool.GetMeasurement(((short) m.QuantityInBaseUnits >> bitPosition) & 1, 0, Measurement.UNITLESS)));
+                                         MeasurementPool.GetMeasurement(((long) m.QuantityInBaseUnits >> bitPosition) & 1, 0, Measurement.UNITLESS)));
                 ed.PushInputData(this, data.DataWithStreamConfiguration(this, this.Configuration));
             }
         }
